# docs/swagger.yaml
basePath: /
definitions:
  dto.AddSubFromWeb:
    properties:
      month:
        example: 5
        type: integer
      price:
        example: 500
        type: integer
      service_name:
        example: YandexGold
        type: string
      start_date:
        example: 02-2022
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  dto.UpdateSubFromWeb:
    properties:
      id:
        example: 1
        type: integer
      month:
        example: 5
        type: integer
      price:
        example: 500
        type: integer
      service_name:
        example: YandexGold
        type: string
      start_date:
        example: 02-2022
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  web.Response:
    properties:
      data: {}
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing subscriptions
  title: Subscription Service API
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: Returns Hello World message
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/web.Response'
      summary: Test endpoint
      tags:
      - Test
  /add_sub:
    post:
      consumes:
      - application/json
      description: Add a new subscription
      parameters:
      - description: Subscription data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddSubFromWeb'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.Response'
      summary: Add subscription
      tags:
      - Subscriptions
  /delete_sub/{id}:
    delete:
      consumes:
      - application/json
      description: Delete subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.Response'
      summary: Delete subscription
      tags:
      - Subscriptions
  /get_list:
    get:
      consumes:
      - application/json
      description: Get list of all subscriptions
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.Response'
      summary: Get all subscriptions
      tags:
      - Subscriptions
  /get_list_by_user/{uuid}:
    get:
      consumes:
      - application/json
      description: Get list of subscriptions by user
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.Response'
      summary: Get user subscriptions
      tags:
      - Subscriptions
  /get_price_subs:
    get:
      consumes:
      - application/json
      description: Get subscription price based on filter criteria
      parameters:
      - description: Service name
        in: query
        name: serv
        required: true
        type: string
      - description: User UUID
        in: query
        name: uuid
        required: true
        type: string
      - description: Start date (MM-YYYY)
        in: query
        name: sdate
        required: true
        type: string
      - description: End date (MM-YYYY)
        in: query
        name: edate
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.Response'
      summary: Get subscription price by filter
      tags:
      - Subscriptions
  /get_sub_by_id/{id}:
    get:
      consumes:
      - application/json
      description: Get subscription details by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.Response'
      summary: Get subscription by ID
      tags:
      - Subscriptions
  /update_sub:
    patch:
      consumes:
      - application/json
      description: Update existing subscription
      parameters:
      - description: Subscription data to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubFromWeb'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update subscription
      tags:
      - Subscriptions
schemes:
- http
swagger: "2.0"